backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
openai==1.51.2
python-dotenv==1.0.0
pandas==2.0.3
openpyxl==3.1.2
PyPDF2==3.0.1
pdfplumber==0.11.7
python-docx==1.2.0
Pillow>=10.0.0
pdf2image>=1.17.0
docx2pdf>=0.1.8
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
